Assignment 5
Added exception handling on the file interaction (database input) and the user input for username. For database input, it was important to ensure that the file existed, which Scanner has a built in exception for that. You just had to catch it and handle it. For the client side of this, I had to ensure that they wont input a blank, whitespace only, or null username as that's a bad time.

Other than that, I made some changes in the actual API, mostly for readability. Variables were renamed appropriately and comments were left on confusing sections of code.