Coordination exception handling (#39)
* Coordination Exceptions and Tests

Added Exceptions to the Coordination component and added tests for it, all tests seem to work and improved code to be easier to read and implement in the future.

* Update CoordinationComponentIntegrationTest.java

* Fixed Error

Wild card import in Coordination Integration Test and Coordination Check, fixed this.

* Oops

oopsie forgot to add assertFalse import

* Fixed Typo

Erro instead of Error lol