bring code into the harness branch (#46)
* Multithreaded network (#43)


* Dynamically allocating threads

I found this nifty function within Stream, which has built in multithreading for file reading. Each line is read by a thread, to make sure that they are not stepping over each other.

* Update CoordinationComponent.java (#44)

* Update CoordinationComponent.java

Added thread protection and some implementation to the coordination component.

* Update CoordinationComponent.java

* fixed tests and coordination component

fixed the tests to also be multi-threaded and finished the coordination component to be multi-threadable

* Edited comments

* Coordination component threading (#45)

* Update CoordinationComponent.java

Added thread protection and some implementation to the coordination component.

* Update CoordinationComponent.java

* fixed tests and coordination component

fixed the tests to also be multi-threaded and finished the coordination component to be multi-threadable

* Edited comments

* added userIds

added userIds for multi test to pass

* Update CoordinationComponent.java

Fixed line 53 error

* Update CoordinationComponent.java

changed one comment but to mainly check if githud code editing messed up

* cheskstyle fix

checkstyle fixes, if this doesnt work with git a new branch will be made

---------

Co-authored-by: ImEvwan <61602941+ImEvwan@users.noreply.github.com>
Co-authored-by: EdwinYang123 <yange3@newpaltz.edu>